fix property order and xml mapping #EA-4079
