+++ entity-management-definitions/src/main/java/eu/europeana/entitymanagement/definitions/model/Aggregation.java
@@ -111,7 +111,7 @@ public class Aggregation {
         + id
         + '\''
         + ", type='"
-        + type
+        + getAggregationType()
         + '\''
         + ", rights='"
         + rights
